+++ CHANGELOG.md
@@ -36,3 +36,10 @@
 1. As Email Address Subscriber concern
 2. Improved the As Contact concern to add the #is_subscribed_as_contact! method
 3. Improved the As Subscriber concern to add the #subscribe_contact! method
+
+## v1.3
+1. Improve the As Contact concern to add the #is_subscribed_as_contact? method
+2. Improve the As Contact concern for the #is_subscribed_as_contact! method
+3. Improve the As Subscriber concern to add the #subscribe_contact? method
+4. Improve the As Subscriber concern for the #subscribe_contact! method
+5. Improve the Contact Subscription model for the .subscribe! method